% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageBh.R
\name{ageBh}
\alias{ageBh}
\title{Calculate age at breast height from site index, top height and species}
\usage{
ageBh(H40, Ho, SP)
}
\arguments{
\item{H40}{site index}

\item{Ho}{top height, i.e. mean o the 100 largest trees with respect to dbh per hectar (m)}

\item{SP}{species code (1=spruce, 2= pine, 3= birch)}
}
\value{
age at breast height
}
\description{
Calculate age at breast height from site index, top height and species
}
\details{
reference age of 40 years
}
\examples{
H40 <- c(11,14,17,20,23) 
Ho <- c(13.995,11.169,21.267,24.81046,28.32)
ageBh( H40 , Ho, 1 )
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@gmail.com} 
modified by Maria Aasnes Moan \email{maria.asnes.moan@nmbu.no}
}
