% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basisHeight.R
\name{basisHeight}
\alias{basisHeight}
\title{basisHeight}
\usage{
basisHeight(d, sp)
}
\arguments{
\item{d}{dbh in cm. Might be a vector.}

\item{sp}{species. 1 = spruce, 2 = pine, 3 = decideous. Should be of same length as h.}
}
\value{
a vector of the same length as the input values.
}
\description{
Calculate basis height as described in Eid, Fitje & Hoen 2002.
}
\author{
Marius Hauglin 2016 \email{marius.hauglin@nmbu.no}
}
\references{
Eid, T.,Fitje A. og Hoen, H.F. (2002) Økonomi og Planlegging. Gan Forlag, Oslo. 205 p.
}

