% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassTree.R
\name{biomassTree}
\alias{biomassTree}
\title{BIOMASS FUNCTIONS}
\arguments{
\item{d}{Numerical vector with diameter at breast height (cm)}

\item{h}{Numerical vector with tree heights (m)}

\item{sp}{Numerical vector with tree species (1 = Norway spruce, 2 = Scots Pine, 3 = Birch)}

\item{components}{String vector specifying which biomass components to return:

 \itemize{
 \item \bold{sb}  Biomass of stem bark
 \item \bold{sw}  Stem wood biomass
 \item \bold{st}  Total stem biomass (independent model including sw and sb)
 \item \bold{fl}  Foliage biomass 
 \item \bold{cr}  Crown biomass 
 \item \bold{br}  Branch biomass (sum of predictions of cr-fl)
 \item \bold{db}  Biomass of dead branches
 \item \bold{su}  Stump biomass 
 \item \bold{rf}  Biomass of fine roots 
 \item \bold{rc}  Biomass of coarse roots 
 \item \bold{rs}  Biomass of roots (sum of predictions of rc+rf)
 \item \bold{sr}  Biomass of stump-root system 
 \item \bold{ab}  Total aboveground biomass (sum of predictions of sw + sb + cr + su + db)
 \item \bold{bg}  Total belowground biomass (using Petersson & Ståhl 2006) 
 \item \bold{tb}  Total tree biomass (sum of ab + bg) 


 \item \bold{all}  All the above components
}}
}
\value{
A data frame with the dry weight biomass (kg) for the specified components. Column names and order will be the same as in components.
}
\description{
Single-tree biomass functions for spruce, pine and birch.
}
\note{
\code{h} is optional. Functions with diameter as the only expl. variable will be used if h is not given. \cr\cr
According to FAO the stump is included in the above-ground biomass. \bold{su} is subtracted from the below-ground biomass
calculated using Petterson & Ståhl's function, since this initially includes the stump. \cr\cr
\bold{su} and \bold{sr} are for birch calculated using the function for pine. \cr\cr
\bold{fl}, \bold{rf} and \bold{rc} are for birch calculated as \bold{sw}*f, where f is a factor derived from ??
}
\references{
Marklund, L.G., 1988. Biomass functions for pine, 
spruce and birch in Sweden (Report No. 45). Swedish University of Agricultural Sciences, Umeå.

Petersson, H. & Ståhl, G. 2006. Functions for below-ground biomass of Pinus sylvestris, Picea abies, Betula pendula and Betula pubescens in Sweden. 
Scandinavian Journal of Forest Research 21(Suppl 7): 84-93

FAO, 2010. Global Forest Resources Assessement 2010 - Terms and definitions. 
Food and Agriculture Organization of the United Nations Working paper 144/E.
}
\author{
Marius Hauglin & Ole Martin Bollandsås (2013-2017) \email{marius.hauglin@gmail.com}
}
