\name{CalcStandVolume}
\alias{CalcStandVolume}
\title{Bergener bestandsvolum}
\usage{
  CalcStandVolume(G, HL, SP)
}
\arguments{
  \item{V}{= Volume, m3/ha}

  \item{G}{= Basal area, m2/ha}

  \item{HL}{= Mean heigh (Lorey's)}

  \item{SP}{= species (1=spruce, 2= pine, 3= brich)}
}
\value{
  V
}
\description{
  Norwegian stand volume functions
}
\details{
  ?
}
\note{
  May 2, 2011
}
\author{
  Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
\references{
  ?
}

