% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corFactor.R
\name{corFactor}
\alias{corFactor}
\title{corFactor}
\usage{
corFactor(sp, cf)
}
\arguments{
\item{sp}{input species: 1 = spruce, 2 = pine, 3 = deciduous}

\item{cf}{correction factor}
}
\value{
a vector of length three with the correction factor for the three species (spruce, pine, deciduous).
}
\description{
Calculate correction factor by converting the correction factor from another species.
}
\note{
the input correction factor is also returned.
}
\author{
Marius Hauglin 2017 \email{marius.hauglin@nmbu.no}
}
\references{
Eid, T. & Fitje, A. 1993. Variasjoner innen bestand for volum, grunnflate, treantall, middeldiameter og middelhøyde.
Meddelelser fra Skogforsk 46.10: 44 s.
}

