% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameterIncrementStand.R
\name{diameterIncrementStand}
\alias{diameterIncrementStand}
\title{Annual diameter increment with bark}
\arguments{
\item{SP}{species code (1=spruce, 2= pine, 3= brich)}

\item{Ho}{Topheight (m)}

\item{H40}{Site index}

\item{N2}{Number of stems after thining}

\item{T13}{Age}

\item{Dmb}{Diamter under bark}
}
\value{
Annual diameter increment with bark (mm)
}
\description{
THIS FUNCTION CALCULATES ANNUAL DIAMETER INCREMENT IN MM WITH BARK FOR THE NORWEGIAN SPECIES 
SPRUCE, PINE, BIRCH AND ASPEN (GRAN, FURU, BJERK ELLER OSP)
}
\details{
TODO: Need to check if function is correct.
}
\examples{
diameterIncrementStand(1,17,150,20)
diameterIncrementStand(1,17,150,20,20,80)
}
\references{
Blingsmo, K. R. 1984. Diametertilvekstfunksjoner for bjork-,furu- og granbestand. NISK-rapport 7/84.
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
