% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameterIncrementTree.R
\name{diameterIncrementTree}
\alias{diameterIncrementTree}
\title{Calculates: Five-year diameter increment of a single tree (mm per 5 yr).}
\arguments{
\item{speciesCode}{species code (1=spruce, 2= pine, 3= brich)}

\item{d}{Diameter of the single tree at breast height (cm)}

\item{BA}{Stand basal area (m2ha-1)}

\item{BAL}{Sum grunnflate på trær større enn treet man beregner for.}

\item{SI}{Site index of spruce (H40-system m.)}

\item{region}{Dummy variable for region. =1 --> County nr >=10}

\item{LAT}{latitude}
}
\value{
V Volume including bark (m3/ha)
}
\description{
Ole Martin Bollandsås 2007. Manuscript USED IN MATRIX MODEL
}
\details{
?
}
\examples{
load(file="data/skogRtrees.rda")
attach(skogRtrees)
#Compute BAL
BAL <- c()
for(i in c(1:length(ba))){
     BAL[i] <- sum(ba[ba>ba[i]])
}
BA <- sum(ba)*(10000/250)
region = 5
LAT = 60.8
SI = 17
diameterIncrementTree(sp,d,BAL,SI,BA,region,LAT)
}
\references{
Vestjordet 1959,Brantseg 1959,Næsset 1994a, Næsset 1995a,b
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
