% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/domHeight.R
\name{domHeight}
\alias{domHeight}
\title{domHeight}
\usage{
domHeight(h, d, area)
}
\arguments{
\item{h}{numerical vector; heights in m.}

\item{d}{numerical vector; diameters in cm.}

\item{area}{numerical; area from which the trees dsecribed by d and h are taken. Given in m2.}
}
\value{
the dominant height in m.
}
\description{
Calculate dominant height
}
\author{
Marius Hauglin 2016 \email{marius.hauglin@nmbu.no}
}
\references{
Fitje, A. (1989). Tremåling.
}

