% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fieldPlot.R
\name{fieldPlot}
\alias{fieldPlot}
\title{fieldPlot}
\usage{
fieldPlot(plotid, d, sp, h, plot_radius, correction = FALSE,
  rel_factor = NA, outside_plot = NA, biomass_components = NA,
  separate_sampletrees = FALSE, corf_method = 2, min_samtr = 3,
  output_heights = FALSE)
}
\arguments{
\item{plotid}{vector giving the plotid for each tree}

\item{d}{numerical vector; diameters in cm.}

\item{sp}{numerical vector; species. (1 = spruce, 2 = pine, 3 = deciduous).}

\item{h}{numerical vector; heights in m.}

\item{plot_radius}{numerical; the plot radius in m.}

\item{correction}{logical; Should be TRUE if sample trees are selected with relascope but from only inside the plot.}

\item{rel_factor}{numerical; if correction is TRUE a relascope factor must be given.}

\item{outside_plot}{logical vector; indicating if a tree is outside the plot. Optional.}

\item{biomass_components}{character vector; calculate plotwise biomass for the given components in Mg/ha. See help
in function biomassTree.
Trees outside the plot will only be used to calculate the correction factor.}

\item{separate_sampletrees}{logical; Determines if sample trees are only used for calculation of correction factor.}

\item{corf_method}{numeric; Method to use if there are missing sample trees for one or more species on a sample plot:
(1) Convert the correction factor from one of the other species on the plot, using the corFactor function. In this alternative no data
from other plots are used.
(2) Use the mean correction factor for the species for all plots.}

\item{min_samtr}{numeric; Minimum number of sample trees allowed for each species in one plot. If the number of sample trees for one species
is below this number then the mean correction factor for all plots are used instead of the local correction factor.}

\item{output_heights}{logical; Estimated heights are returned for all trees.}
}
\value{
a data frame.
if output_heights is TRUE: a list with two items, a data.frame with plotwise results, and a data.frame with the
estimated heights for all trees in the dataset, in the same order as the input.
}
\description{
Calculates volume, basal area, stem number and mean and dominant height on field plots
}
\author{
Marius Hauglin 2017 \email{marius.hauglin@gmail.com}
}
\references{
Eid, T., Fitje, A., and Hoen, H.F. (2002) Økonomi og planlegging. Gan Forlag, Oslo. 205 s.

Bollandsås, O.M. (2007). Calculating volume on the field plots of the Norwegian National Forest Inventory. Unpublished.
}

