% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fraStikkliste.R
\name{fraStikkliste}
\alias{fraStikkliste}
\title{fraStikkliste}
\usage{
fraStikkliste(id, sp, stikkliste, classes = NA)
}
\arguments{
\item{id}{vector of ids}

\item{sp}{numerical vector; species: 1 = spruce, 2 = pine, 3 = deciduous}

\item{stikkliste}{data.frame or matrix; number of trees in each diameter class. The columns are the diameter classes in increasing order.
Number of rows equal to the length of id and sp vectors.}

\item{classes}{numerical vector; diameter for each diameter class. Length must equal number of columns in stikkliste.}
}
\value{
a data.frame with the columns: id, species and dbh (diameter).
}
\description{
creates table of trees from a 'stikkliste'
}
\author{
Marius Hauglin 2017 \email{marius.hauglin@gmail.com}
}

