% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeightGrowth.R
\name{HeightGrowth}
\alias{HeightGrowth}
\title{HeightGrowth}
\usage{
HeightGrowth(Hdt1, H40, years, ts)
}
\arguments{
\item{Hdt1}{a numeric vector of initial dominant heights (m)}

\item{H40}{site index (m)}

\item{years}{observation period (nr of years between time points 1 and 2)}

\item{ts}{species code (1=spruce, 2= pine, 3= birch)}
}
\value{
Expected dominant heights at time point 2
}
\description{
Function for future dominant height prognoses
}
\details{
Reference age of 40 years
}
\examples{
a <- c(8,20) # initial dominant heights
b <-c(8,20) # H40 site index values
HeightGrowth(Hdt1=a, H40=b, years=15, ts=1)
}
\references{
Eriksson (1997), Sharma et al (2011)
}
\seealso{
SiteIndex, HeightDiffSi
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@gmail.com}
}
