% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightRelativGrowthTree.R
\name{heightRelativGrowthTree}
\alias{heightRelativGrowthTree}
\title{Calculates relative height growth of single trees}
\usage{
heightRelativGrowthTree(d, sp, CIndex = "CI4")
}
\arguments{
\item{d}{Diameter of the single tree at breast height (cm)}

\item{sp}{species code (1=spruce, 2= pine, 3= brich)}

\item{CIndex}{competition index to be used only "CI4" implemented}
}
\value{
Height growth for the period (m)
}
\description{
Calculate relative height growth of singler trees based on Sharma et al. 2017
}
\details{
?
}
\examples{
sp <- c(1,1,2,4)
d <- c(20,25,10,5)
heightRelativGrowthTree(d,sp)
}
\references{
Sharma, R.P. & Brunner, A., 2017. Modeling individual tree height growth of Norway spruce and Scots pine from national forest inventory data in Norway. Scandinavian journal of forest research / issued bimonthly by the Nordic Forest Research Cooperation Committee , 32(6), pp.501–514.
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
