% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightStandDominantDevelopment.R
\name{heightStandDominantDevelopment}
\alias{heightStandDominantDevelopment}
\title{Dominant height development (Sharma et al. 2011)}
\arguments{
\item{H40}{Site index of spruce (H40-system m.)}

\item{AGE}{Stand age}

\item{DominantSpecies}{Dominant species of the stand (species code: 1=spruce, 2= pine)}
}
\value{
Dominant hegith (m)
}
\description{
Calculates dominant height for Norway spruce (Picea abies (L.) Karst.) and Scots pine (Pinus sylvestris L.) in Norway.
}
\details{
See reference
}
\examples{
dominantHeightDevelopment(70,17,1)
dominantHeightDevelopment(70,17,2)
dominantHeightDevelopment(70,17,3)
}
\references{
Sharma, R.P. et al., 2011. Modelling dominant height growth from national forest inventory individual tree data with short time series and large age errors. Forest ecology and management, 262(12), pp.2162–2175.
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
