% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTree.R
\name{mortalityTree}
\alias{mortalityTree}
\title{Calculates 5 year mortality rate for tree-species}
\arguments{
\item{species}{species code (1=spruce, 2= pine, 3= brich, 4 = other broadleaves )}

\item{d}{Diameter of the single tree at breast height (cm), Diameter (cm)(Publication uses cm, parameter estimate adjusted)}

\item{BA}{Stand basal area (m2ha-1)}
}
\value{
5 year mortality rate
}
\description{
#' 'Model for survival. Original model by Eid & Tuhus (2005). Later (april 2007) the model was modified to a more simple model and to include basal area. Eid, T. & Tuhus, E., (2005). Models for individual tree mortality in Norway. For. Ecol. Manage. 154: 69-84
}
\details{
Based on SAS code from 'Ole Martin Bollandsås, 16.01.2007
}
\examples{
sp <- c(1,1,2,4)
d <- c(20,25,10,5)
BA <- 14
mortalityTree(d,sp,BA)
BA <- c(14,14,10,10)
mortalityTree(d,sp,BA)
}
\references{
Vestjordet 1959,Brantseg 1959,Næsset 1994a, Næsset 1995a,b
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
