% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteIndexMultiTemp.R
\name{siteIndexMultiTemp}
\alias{siteIndexMultiTemp}
\title{siteIndexMultiTemp}
\usage{
siteIndexMultiTemp(Ho, yrs, SP)
}
\arguments{
\item{Ho}{values of top heights in m}

\item{yrs}{numeric values of years of measurement (e.g. 2020)}

\item{SP}{species code (1=spruce, 2= pine, 3= birch)}
}
\value{
list of three elements: (1) H40 site index (m), (2) calculated ages for the input values of height, and (3) absolute residual (m)
}
\description{
Calculate site index from time series of top heights and years of measurement
}
\details{
Reference age of 40 years
}
\examples{
Ho = c(17, 20, 23)
yrs = c(1999, 2010, 2022)
SP = 1
siteIndexMultiTemp( Ho , yrs , SP )
}
\references{
Sharma et al (2011)
}
\seealso{
HeightGrowth, HeightDiffSi, SiteIndex
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@gmail.com}
}
