% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skogR.News}
\alias{skogR.News}
\title{Show the NEWS file.}
\usage{
skogR.News()
}
\value{
None
}
\description{
Show the NEWS file of the NorForFunc package.
}
\examples{
#skogR.News()
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}

