% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/volumeDoubleBark.R
\name{volumeDoubleBark}
\alias{volumeDoubleBark}
\title{Helpfunctions for volume: Doublebarkthickness}
\usage{
volumeDoubleBark(d, h, sp, Fitje1995 = FALSE)
}
\arguments{
\item{d}{Diameter in breast hegiht i centimeter including bark}

\item{h}{Tree height in meter}

\item{sp}{species codes (1=spruce, 2=pine, 3=Birch)}

\item{Fitje1995}{logical indicating to use the functions by Fitje 1995 rather than Brantseg 1967}
}
\value{
b doublebark i mm
}
\description{
Functions for calculation of doublebark in milimeter
}
\details{
bla bla
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
\references{
Braastad, H. 1966. Volume tables for birch. Reports of the Norwegian Forest Research Institute, 21, 265-365.
Brantseg, A. 1967. Volume functions and tables for Scots pine. South Norway. Reports of the Norwegian Forest Research Institute, 12, 689-739.
Vestjordet, E. 1967. Functions and tables for volume of standing trees. Norway spruce. Reports of the Norwegian Forest Research Institute, 22, 539-574.
Fitje A. 1995.
}

