% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumeStand.R
\name{volumeStand}
\alias{volumeStand}
\title{Stand volume}
\usage{
volumeStand(G, HL, SP, Vestlandet = NULL)
}
\arguments{
\item{G}{Basal area (m2/ha)}

\item{HL}{Lorey's Mean heigh (m)}

\item{SP}{species code (1=spruce, 2= pine, 3= brich)}

\item{Vestlandet}{character for using functions representative for western-Norway either "coast" or "interior"}
}
\value{
V Volume including bark (m3/ha)
}
\description{
Compute stand volume based on Norwegian stand volume functions
}
\details{
?
}
\references{
Vestjordet 1959,Brantseg 1959,Næsset 1994a, Næsset 1995a,b
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
