% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biomassTree.R
\name{biomassTree}
\alias{biomassTree}
\title{BIOMASS FUNCTIONS (Marklund, 1988)}
\usage{
biomassTree(d, h = NA, sp, components = c("all"))
}
\arguments{
\item{d}{Numerical vector with diameter at breast height (cm)}

\item{h}{Numerical vector with tree heights (m)}

\item{sp}{Numerical vector with tree species (1 = Norway spruce, 2 = Scots Pine, 3 = Birch)}

\item{components}{String vector specifying which biomass components to return:

 \itemize{
 \item \bold{sb}  Biomass of stem bark
 \item \bold{sw}  Stem wood biomass
 \item \bold{st}  Total stem biomass (sw+sb)
 \item \bold{fl}  Foliage biomass
 \item \bold{cr}  Crown biomass
 \item \bold{br}  Branch biomass (cr-fl)
 \item \bold{db}  Biomass of dead branches
 \item \bold{su}  Stump biomass *
 \item \bold{rf}  Biomass of fine roots *
 \item \bold{rc}  Biomass of coarse roots *
 \item \bold{rs}  Biomass of roots (rc+rf) *
 \item \bold{sr}  Biomass of stump-root system *
 \item \bold{ab}  Total aboveground biomass (st + cr + su + db)
 \item \bold{bg}  Total belowground biomass (Petterson & Ståhl 2006)
 \item \bold{tb}  Total tree biomass (st + cr + db + bg)


 \item \bold{all}  All the above components
    }}
}
\value{
A data frame with the dry weight biomass (kg) for the specified components. Column names and order will be the same as in components.
}
\description{
Single-tree biomass functions for spruce, pine and birch.
}
\note{
\code{h} is optional. Marklund include functions with diameter alone, and with diameter
and height as explanatory variables for most biomass components.
Note that for some components only functions with
diameter as single explanatory variable are available (marked with * in the list above).
For these biomass components the
values from the functions with diameter as single explanatory variable will
be returned for spruce and pine, \emph{even if height values are provided.}
For birch there are no functions for these biomass components
(marked with *) so NAs will be returned.
In the case of the crown biomass component (\bold{cr}) for birch the function with
diameter as single explanatory variable is the only
available function, thus the value from this is returned,
\emph{even if height values are provided.}
}
\author{
Marius Hauglin (2013-2017) \email{marius.hauglin@nmbu.no}
}
\references{
Marklund, L.G., 1988. Biomass functions for pine,
spruce and birch in Sweden (Report No. 45). Swedish University of Agricultural Sciences, Umeå.

Petersson, H. & Ståhl, G. 2006. Functions for below-ground biomass of Pinus sylvestris, Picea abies, Betula pendula and Betula pubescens in Sweden.
Scandinavian Journal of Forest Research 21(Suppl 7): 84-93
}

