% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crownHeightPine.R
\name{crownHeightPine}
\alias{crownHeightPine}
\title{Crown height of pine}
\usage{
crownHeightPine(d, h)
}
\arguments{
\item{d}{Diameter in breast height (cm)}

\item{h}{Tree height (m)}
}
\value{
k crownheight (m)
}
\description{
Compute crown height of single trees using Norwegian only for pine...
}
\references{
Brantseg 1967
}
\seealso{
volumeTreePine2
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
