% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fieldPlot.R
\name{fieldPlot}
\alias{fieldPlot}
\title{fieldPlot}
\usage{
fieldPlot(plotid, d, sp, h, correction = FALSE, rel_factor = NA,
  plot_radius = NA, outside_plot = NA)
}
\arguments{
\item{plotid}{vector giving the plotid for each tree}

\item{d}{numerical vector; diameters in cm.}

\item{sp}{numerical vector; species. (1 = spruce, 2 = pine, 3 = deciduous).}

\item{h}{numerical vector; heights in m.}

\item{correction}{logical; Should be TRUE if sample trees are selected with relascope but from only inside the plot.}

\item{rel_factor}{numerical; if correction is TRUE a relascope factor must be given.}

\item{plot_radius}{numerical; if correction is TRUE a plot radius must be given.}

\item{outside_plot}{logical vector; indicating if a tree is outside the plot. Optional.
Trees outside the plot will only be used to calculate the correction factor.}
}
\value{
a data frame with the following columns:
}
\description{
Calculates volume, basal area, stem number and mean and dominant height on field plots
}
\author{
Marius Hauglin 2017 \email{marius.hauglin@gmail.com}
}
\references{
Eid, T., Fitje, A., and Hoen, H.F. (2002) Økonomi og planlegging. Gan Forlag, Oslo. 205 s.

Bollandsås, O.M. (2007). Calculating volume on the field plots of the Norwegian National Forest Inventory. Unpublished.
}

