% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fieldPlot.R
\name{fieldPlot}
\alias{fieldPlot}
\title{fieldPlot}
\usage{
fieldPlot(plotid, d, sp, h, correction, rel_factor = NA, plot_radius = NA,
  outside_plot = NA, esth_method = 1)
}
\arguments{
\item{plotid}{vector giving the plotid for each tree}

\item{d}{numerical vector; diameters in cm.}

\item{sp}{numerical vector; species.}

\item{h}{numerical vector; heights in m.}

\item{correction}{logical; Should be TRUE if sample trees are selected from only inside the plot.}

\item{rel_factor}{numerical; if correction is TRUE a relascope factor must be given.}

\item{plot_radius}{numerical; if correction is TRUE a plot radius must be given.}

\item{outside_plot}{logical vector; indicating if a tree is outside the plot. Optional.}

\item{esth_method}{numerical; giving the method to use for estimation of tree heights. See details.}
}
\value{
a data frame with the following columns:
}
\description{
Calculates volume, basal area, stem number and mean and dominant height on field plots
}
\note{
Trees outside the plot will be used to calculate the correction factor.
}
\author{
Marius Hauglin 2016 \email{marius.hauglin@nmbu.no}
}
\references{
Eid, T., Fitje, A., and Hoen, H.F. (2002) Økonomi og planlegging. Gan Forlag, Oslo. 205 s.

Bollandsås, O.M. (2007). Calculating volume on the field plots of the Norwegian National Forest Inventory. Unpublished.
}

