% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heightFromVolume.R
\name{heightFromVolume}
\alias{heightFromVolume}
\title{heightFromVolume}
\usage{
heightFromVolume(vol, d, sp, interval = c(0, 45), obj.threshold = NA)
}
\arguments{
\item{vol}{numerical vector; single tree volume. dm3.}

\item{d}{numerical vector; diameters in cm.}

\item{sp}{numerical vector; species.}

\item{interval}{numeric vector; default lower/upper bound for optim}

\item{obj.threshold}{numeric; the function will return NA if the objective value from the optimisation is larger than this value.}
}
\value{
a vector of the same length as the input with estimated heights.
}
\description{
Calculate estimated heights by reversing the single-tree volume functions.
}
\author{
Marius Hauglin 2016 \email{marius.hauglin@gmail.com}
}

