% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightStandDominant.R
\name{heightStandDominant}
\alias{heightStandDominant}
\title{Dominant height}
\usage{
heightStandDominant(H40, age, SP)
}
\arguments{
\item{H40}{a vector of site indeces (m)}

\item{age}{age at breast height (years)}

\item{SP}{species code (1=spruce, 2= pine, 3 = birch, 4= aspen)}
}
\value{
Dominant tree height (m)
}
\description{
Calculates dominant height based on site index and age
}
\details{
Index age is 40
}
\examples{
H40 <- c(11,14,17,20,23)
age <- c(40,40,40,50,60)
heightStandDominant( H40 = H40, age = age , SP = 2 )
}
\references{
Sharma et. al. (2011), Eriksson et. al. (1997), Opdahl (1992)
}
\seealso{
SiteIndex, HeightDiffSi, HeightGrowth, BhAge
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@nmbu.no}
}
