% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteIndex.R
\name{siteIndex}
\alias{siteIndex}
\title{SiteIndex}
\usage{
siteIndex(age, HO, SP, method = "SHARMA-BRUNNER")
}
\arguments{
\item{age}{age at breast height (years)}

\item{HO}{top height, i.e. mean of the 100 largest trees with respect to dbh per hectar (m)}

\item{SP}{species code (1=spruce, 2= pine, 3= birch)}

\item{method}{character for choosing the intended functions, default or "TVEITE-BRAASTAD"}
}
\value{
H40 site index (m)
}
\description{
calculate site index according to the Norwegian H40 system
}
\details{
Reference age of 40 years
}
\examples{
a <- c(55,55) # ages at breast height
b <- c(10.23744, 24.81046) # dominant heights (m)
SiteIndex( a, b, 1) 
}
\references{
Tveite (1977), Tveite & Braastad (1981), Eriksson (1997), Sharma et al (2011)
}
\seealso{
HeightGrowth, HeightDiffSi
}
\author{
Lennart Noordermeer \email{lennart.noordermeer@gmail.com}
}
