% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skogRNews.R
\name{skogR.News}
\alias{skogR.News}
\title{Show the NEWS file.}
\value{
None
}
\description{
Show the NEWS file of the NorForFunc package.
}
\examples{
skogR.News()
}
\author{
Hans Ole Ørka \email{hans.ole.orka@gmail.org}
}
